Move inbound and outbound stacks from app::main
