+++ src/app/errors.rs
@@ -127,9 +127,7 @@ fn map_err_to_5xx(e: Error) -> StatusCode {
     } else if let Some(_) = e.downcast_ref::<router::NotRecognized>() {
         error!("could not recognize request");
         http::StatusCode::BAD_GATEWAY
-    } else if let Some(err) =
-        e.downcast_ref::<outbound::require_identity_on_endpoint::RequireIdentityError>()
-    {
+    } else if let Some(err) = e.downcast_ref::<outbound::RequireIdentityError>() {
         error!("{}", err);
         http::StatusCode::FORBIDDEN
     } else {
