+++ src/app/mod.rs
@@ -13,6 +13,7 @@ mod main;
 mod metric_labels;
 mod outbound;
 mod profiles;
+mod proxy;
 mod tap;
 
 pub use self::main::Main;
@@ -69,3 +70,21 @@ fn http_request_orig_dst_addr<B>(req: &http::Request<B>) -> Result<Addr, addr::E
         .map(Addr::Socket)
         .ok_or(addr::Error::InvalidHost)
 }
+
+#[derive(Copy, Clone, Debug)]
+struct DispatchDeadline(std::time::Instant);
+
+impl DispatchDeadline {
+    fn after(allowance: std::time::Duration) -> DispatchDeadline {
+        DispatchDeadline(tokio_timer::clock::now() + allowance)
+    }
+
+    fn extract<A>(req: &http::Request<A>) -> Option<std::time::Instant> {
+        req.extensions().get::<DispatchDeadline>().map(|d| d.0)
+    }
+}
+
+type HttpEndpointMetricsRegistry =
+    crate::proxy::http::metrics::SharedRegistry<metric_labels::EndpointLabels, classify::Class>;
+type HttpRouteMetricsRegistry =
+    crate::proxy::http::metrics::SharedRegistry<metric_labels::RouteLabels, classify::Class>;
